Fixed price of item

+++ app/src/main/java/cat/udl/menufinder/fragments/ManageItemsFragment.java
@@ -72,7 +72,7 @@ public class ManageItemsFragment extends MasterFragment {
 
                 if (closeDialog) {
                     alertDialog.dismiss();
-                    savetoDB(new Item(name, description, Integer.valueOf(price)));
+                    savetoDB(new Item(name, description, Double.valueOf(price)));
                 }
             }
         });
